solar-panel-detection-main/dataflow/requirements.txt
apache-beam[gcp]==2.53.0
psycopg2-binary==2.9.9
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
pandas==2.1.4
pyarrow==14.0.1
fastparquet==2023.10.1
